refactor: extract admin routes into ui/admin.py
Move 37 admin routes (/admin/*, resolution, venue/event management,
dedup, tasks, stats) to ui/admin.py. All routes converted from inline
auth checks and _get_db context manager to FastAPI dependency injection
(require_admin, get_db). Auth behavior change: unauthenticated requests
now get 307 redirect to /login instead of 403 (consistent with other
DI-converted routes). routes.py reduced from 2992 to 1132 lines.
