+++ src/resonance/app.py
@@ -28,6 +28,7 @@ import resonance.logging as logging_module
 import resonance.middleware.session as session_middleware
 import resonance.models.task as task_models
 import resonance.types as types_module
+import resonance.ui.admin as ui_admin_module
 import resonance.ui.artists as ui_artists_module
 import resonance.ui.events as ui_events_module
 import resonance.ui.routes as ui_routes_module
@@ -102,6 +103,7 @@ def create_app() -> fastapi.FastAPI:
     application.include_router(api_v1_module.router)
 
     # Register UI routes
+    application.include_router(ui_admin_module.router)
     application.include_router(ui_artists_module.router)
     application.include_router(ui_events_module.router)
     application.include_router(ui_tracks_module.router)
